+++ lib/src/element_params/sub/button/popupmenu_btn_params.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:simple_managers/simple_managers.dart';
 
 class PopupMenuBtnParamsWrapper {
   PopupMenuBtnParams p;
@@ -23,7 +24,9 @@ class PopupMenuBtnParams {
   // popupmenu button parameters
   Key? key;
   List<PopupMenuEntry<int>> Function(BuildContext)? itemBuilder;
-  int? initialValue;
+  // pre version, this is initialValue.
+  IndexManager? manager;
+  void Function()? onOpened;
   void Function(int)? onSelected;
   void Function()? onCanceled;
   String? tooltip;
@@ -40,6 +43,7 @@ class PopupMenuBtnParams {
   bool? enableFeedback;
   BoxConstraints? constraints;
   PopupMenuPosition position = PopupMenuPosition.over;
+  Clip clipBehavior = Clip.none;
   // others
   List<PopupMenuItemParams> popupMenuItemParams = [];
   void Function(int index)? menuCallback;