+++ lib/src/element_params/sub/button/segmented_btn_params.dart
@@ -1,6 +1,7 @@
 import 'dart:core';
 
 import 'package:flutter/material.dart';
+import 'package:simple_managers/simple_managers.dart';
 
 class SegmentedBtnParamsWrapper {
   SegmentedBtnParams p;
@@ -14,7 +15,8 @@ class SegmentedBtnParams {
   static const defEmptySelectionAllowed = true;
   Key? key;
   List<ButtonSegment<int>>? segments;
-  Set<int>? selected;
+  // pre version, This is selected params.
+  MultiIndexManager? manager;
   // The set value is children indexes.
   void Function(Set<int>)? onSelectionChanged;
   bool multiSelectionEnabled = defMultiSelectionEnabled;